% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_reorder.R
\name{desc_reorder}
\alias{desc_reorder}
\title{Reorder levels of a factor according to some corresponding numeric values in descending order}
\usage{
desc_reorder(x, y, fun = mean)
}
\arguments{
\item{x}{A vector of class factor to be reordered}

\item{y}{A numeric vector}

\item{fun}{a function whose first argument is a vector and returns a scalar, to be applied
to each subset of x determined by the levels of y}
}
\value{
A factor with desired reordered levels
}
\description{
Reorder levels of a factor according to some corresponding numeric values in descending order
}
\examples{
boxplot(Sepal.Width ~ Species, data = iris)
boxplot(Sepal.Width ~ desc_reorder(Species, Sepal.Width), data = iris)
}

