% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_char.R
\name{detect_char}
\alias{detect_char}
\title{Detect a vector that might be of class character}
\usage{
detect_char(x, convert = FALSE)
}
\arguments{
\item{x}{a vector}

\item{convert}{a logical value indicating whether to convert the vector into character if it is
suspected by the function to be of class character. Default is F}
}
\value{
If convert is T, return converted vector of class character. Otherwise, return nothing but
        print corresponding messages
}
\description{
Detect a vector that might be of class character
}
\examples{
x <- as.factor(c("This", "is", "actually", "a", "sentence"))
class(x)
detect_char(x)
xnew <- detect_char(x, convert = TRUE)
class(xnew)
}

